requirements.txt
streamlit
pypdf
python-dotenv
strands-agents
boto3
botocore
